refactor(dashboard): remove redundant defensive checks
Use the canonical ITSM feature flag and centralize MCP prompt payload validation in the app layer to prevent duplicated behavior while preserving dependency boundaries.
