+++ src/dashboard/apigateway/apigateway/apps/mcp_server/management/commands/fix_invalid_mcp_prompts.py
@@ -24,6 +24,7 @@ from django.utils import timezone
 
 from apigateway.apps.mcp_server.constants import MCPServerExtendTypeEnum
 from apigateway.apps.mcp_server.models import MCPServerExtend
+from apigateway.apps.mcp_server.validators import validate_mcp_prompts_payload
 
 
 def _parse_mcp_server_ids(value: str) -> List[int]:
@@ -41,15 +42,6 @@ def _parse_mcp_server_ids(value: str) -> List[int]:
         raise CommandError("No valid mcp_server_ids provided.")
 
     return ids
-
-
-def _validate_prompts_payload(prompts):
-    if not isinstance(prompts, list):
-        raise TypeError("prompts must be a list")
-
-    for prompt in prompts:
-        if not isinstance(prompt, dict):
-            raise TypeError("prompt item must be a dict")
 
 
 class Command(BaseCommand):
@@ -119,7 +111,7 @@ class Command(BaseCommand):
         for extend in queryset.iterator(chunk_size=200):
             try:
                 prompts = json.loads(extend.content)
-                _validate_prompts_payload(prompts)
+                validate_mcp_prompts_payload(prompts)
             except json.JSONDecodeError as err:
                 invalid_extends.append(extend)
                 self.stdout.write(
