+++ src/dashboard/apigateway/apigateway/apps/mcp_server/validators.py
@@ -5,7 +5,7 @@
 # Licensed under the MIT License (the "License"); you may not use this file except
 # in compliance with the License. You may obtain a copy of the License at
 #
-#     http://opensource.org/licenses/MIT
+# http://opensource.org/licenses/MIT
 #
 # Unless required by applicable law or agreed to in writing, software distributed under
 # the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
@@ -15,25 +15,15 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
-from .mcp_server import (
-    build_mcp_server_application_url,
-    build_mcp_server_detail_url,
-    build_mcp_server_permission_approval_url,
-    build_mcp_server_url,
-    build_mcp_streamable_http_url,
-    update_stage_mcp_server_related_resource_names,
-)
 
-__all__ = [
-    # constant
-    # Enum
-    # class
-    # functions
-    "build_mcp_server_application_url",
-    "build_mcp_server_detail_url",
-    "build_mcp_server_permission_approval_url",
-    "build_mcp_server_url",
-    "build_mcp_streamable_http_url",
-    "update_stage_mcp_server_related_resource_names",
-    # others
-]
+from typing import Any
+
+
+def validate_mcp_prompts_payload(prompts: Any) -> None:
+    """Validate MCP prompts payload shape."""
+    if not isinstance(prompts, list):
+        raise TypeError("prompts must be a list")
+
+    for prompt in prompts:
+        if not isinstance(prompt, dict):
+            raise TypeError("prompt item must be a dict")
